Invest/discord_test/requirements.txt
schedule
discord.py